Add Python "go-between" script
This adds a slightly modified version of the Python "go-between"
script that connects SMAC and Clojush. We cleaned up a few bits
of historical cruft and changed the Clojush problem into a command
line argument so the script won't have to be changed when you go
from one problem to another.

There is still quite a lot of refactoring that can be done here.
The handling of command line arguments is pretty atrocious, for
example, and should definitely be improved.