Report CPU time back to SMAC
In order to make the results more uniform across machines and
circumstances, we want to switch from using wall clock limits to
using CPU time limits. This required actually getting the CPU time
of the Clojush run from `runsolver` and including that in the
results reported back to SMAC.

One consequence of having done this is that we can now use RUNTIME
as the objective instead of always using QUALITY. I'm not sure what
the implications of that will be, but it would be interesting to
try it out.